Implement LSIF emitter in Java.
Previously, we shelled out to `lsif-semanticdb` to convert SemanticDB
files into LSIF. This commit ports the original Go implemementation to
Java so that we can call a normal method instead of shelling out to a
separate process.

The biggest problem with the previous solution is that it complicated
installation for users. As we evolve lsif-java, the conversion logic
will also evolve and we want to avoid the situation where users have
an outdated version of lsif-semanticdb installed. With this commit,
users only need to install `lsif-java` and everything should work
correctly.

This commit also adds a new `snapshot-lsif` command to pretty-print an
LSIF index, similar to how we pretty-print SemanticDB payloads. This
command is mostly useful for testing purposes, but could be helpful
to troubleshoot user issues.